Improved Security / Randomize Data
With this update I require all stored data to be effectively random
according to the chi-sq algorithm.   This allows nodes to verify that
the chunks they are storing are indeed 'random' and not some unencrypted
contraband that could potentially get the host in trouble.  The host can
validate the 'encryption' without having to know anything else about the
file.

I then updated the file import tool to ensure that every chunk is
'random' by the chi-sq algorithm.  Blowfish by itself cannot ensure
that every chunk is 'random' by the chi-sq method, because even 'random'
data can appear 'non-random' some times.  Therefore, I do a brute force
search for a random seed that when combined with the MTwister algorithm
generates a sequence of bytes that when xored with the output of
blowfish encryption generates a sequence of bytes that are 'random' by
the chi-sq measure.

This seed is then stored as part of the tornet file to allow the user
to decrypt the data.